Fix: Fixed the issue where switching between two operator models would cause the classification node connection to be lost
