+++ web/src/pages/flow/flow-drawer/index.tsx
@@ -2,7 +2,7 @@ import { useTranslate } from '@/hooks/common-hooks';
 import { IModalProps } from '@/interfaces/common';
 import { CloseOutlined } from '@ant-design/icons';
 import { Drawer, Flex, Form, Input } from 'antd';
-import { lowerFirst } from 'lodash';
+import { get, isPlainObject, lowerFirst } from 'lodash';
 import { Play } from 'lucide-react';
 import { useEffect, useRef } from 'react';
 import { BeginId, Operator, operatorMap } from '../constant';
@@ -40,7 +40,11 @@ import WikipediaForm from '../form/wikipedia-form';
 import YahooFinanceForm from '../form/yahoo-finance-form';
 import { useHandleFormValuesChange, useHandleNodeNameChange } from '../hooks';
 import OperatorIcon from '../operator-icon';
-import { getDrawerWidth, needsSingleStepDebugging } from '../utils';
+import {
+  buildCategorizeListFromObject,
+  getDrawerWidth,
+  needsSingleStepDebugging,
+} from '../utils';
 import SingleDebugDrawer from './single-debug-drawer';
 
 import { RAGFlowNodeType } from '@/interfaces/database/flow';
@@ -123,10 +127,21 @@ const FormDrawer = ({
       if (node?.id !== previousId.current) {
         form.resetFields();
       }
-      form.setFieldsValue(node?.data?.form);
+
+      if (operatorName === Operator.Categorize) {
+        const items = buildCategorizeListFromObject(
+          get(node, 'data.form.category_description', {}),
+        );
+        const formData = node?.data?.form;
+        if (isPlainObject(formData)) {
+          form.setFieldsValue({ ...formData, items });
+        }
+      } else {
+        form.setFieldsValue(node?.data?.form);
+      }
       previousId.current = node?.id;
     }
-  }, [visible, form, node?.data?.form, node?.id]);
+  }, [visible, form, node?.data?.form, node?.id, node, operatorName]);
 
   return (
     <Drawer
