+++ web/src/pages/flow/form/categorize-form/hooks.ts
@@ -1,38 +1,10 @@
-import get from 'lodash/get';
-import omit from 'lodash/omit';
-import { useCallback, useEffect } from 'react';
 import {
   ICategorizeItem,
   ICategorizeItemResult,
-  IOperatorForm,
-} from '../../interface';
-import useGraphStore from '../../store';
-
-/**
-   * convert the following object into a list
-   * 
-   * {
-      "product_related": {
-      "description": "The question is about product usage, appearance and how it works.",
-      "examples": "Why it always beaming?\nHow to install it onto the wall?\nIt leaks, what to do?",
-      "to": "generate:0"
-      }
-      }
-*/
-const buildCategorizeListFromObject = (
-  categorizeItem: ICategorizeItemResult,
-) => {
-  // Categorize's to field has two data sources, with edges as the data source.
-  // Changes in the edge or to field need to be synchronized to the form field.
-  return Object.keys(categorizeItem)
-    .reduce<Array<ICategorizeItem>>((pre, cur) => {
-      // synchronize edge data to the to field
-
-      pre.push({ name: cur, ...categorizeItem[cur] });
-      return pre;
-    }, [])
-    .sort((a, b) => a.index - b.index);
-};
+} from '@/interfaces/database/flow';
+import omit from 'lodash/omit';
+import { useCallback } from 'react';
+import { IOperatorForm } from '../../interface';
 
 /**
    * Convert the list in the following form into an object
@@ -58,12 +30,7 @@ const buildCategorizeObjectFromList = (list: Array<ICategorizeItem>) => {
 
 export const useHandleFormValuesChange = ({
   onValuesChange,
-  form,
-  nodeId,
 }: IOperatorForm) => {
-  const getNode = useGraphStore((state) => state.getNode);
-  const node = getNode(nodeId);
-
   const handleValuesChange = useCallback(
     (changedValues: any, values: any) => {
       onValuesChange?.(changedValues, {
@@ -73,15 +40,6 @@ export const useHandleFormValuesChange = ({
     },
     [onValuesChange],
   );
-
-  useEffect(() => {
-    const items = buildCategorizeListFromObject(
-      get(node, 'data.form.category_description', {}),
-    );
-    form?.setFieldsValue({
-      items,
-    });
-  }, [form, node]);
 
   return { handleValuesChange };
 };
